# chatgpt-app/openapi.yaml
openapi: "3.1.0"
info:
  title: Course Companion FTE API
  description: >
    Zero-Backend-LLM API for the AI Agent Development course by Panaversity.
    All intelligent reasoning is handled by the ChatGPT App (frontend LLM).
    This backend is 100% deterministic — no LLM calls.
  version: "1.0"
servers:
  - url: https://your-backend.railway.app
    description: Production server (update after Railway deployment)

# ─────────────────────────────────────────────────────────────────────────────
# PATHS
# ─────────────────────────────────────────────────────────────────────────────
paths:
...
  /api/v1/quiz/chapter/{chapter_id}:
    get:
      operationId: get_quiz_id_for_chapter
      summary: Get quiz ID for a chapter
      description: >
        Returns the quiz_id that corresponds to a given chapter_id.
        Use this to find which quiz to load after a student finishes a chapter.
      parameters:
        - name: chapter_id
          in: path
          required: true
          description: Chapter identifier
          schema:
            type: string
            example: chapter-01
      responses:
        "200":
          description: Quiz ID for the chapter
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QuizIdResponse"
        "404":
          description: No quiz found for this chapter
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorDetail"
  /api/v1/quiz/{quiz_id}:
    get:
      operationId: get_quiz
      summary: Get quiz questions (without answer keys)
      description: >
        Returns quiz questions with multiple-choice options A/B/C/D.
        IMPORTANT: Correct answers are NOT included in this response.
        Never guess or reveal answers — always submit to submit_quiz for grading.
      parameters:
        - name: quiz_id
          in: path
          required: true
          description: Quiz identifier (e.g. quiz-01)
          schema:
            type: string
            example: quiz-01
      responses:
        "200":
          description: Quiz questions without answer keys
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QuizOut"
        "404":
          description: Quiz not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorDetail"
  /api/v1/quiz/{quiz_id}/submit:
    post:
      operationId: submit_quiz
      summary: Submit quiz answers for grading
      description: >
        Submits student answers for deterministic rule-based grading.
        ZERO LLM calls — grading is pure string comparison against the answer key.
        Returns score (0-100), pass/fail, and per-question feedback with explanations
        for incorrect answers. Stores attempt in quiz_attempts table.
      parameters:
        - name: quiz_id
          in: path
          required: true
          description: Quiz identifier
          schema:
            type: string
            example: quiz-01
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/QuizSubmission"
      responses:
        "200":
          description: Grading result with score, pass/fail, and per-question feedback
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QuizResult"
        "404":
          description: Quiz not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorDetail"

  # ── 6.4 Progress APIs ────────────────────────────────────────────────────────
  /api/v1/progress/{user_id}:
    get:
      operationId: get_progress
      summary: Get full user progress summary
      description: >
        Returns complete progress data for a user: completed chapters,
        overall completion percentage, current streak, longest streak,
        total study days, and best quiz score per quiz.
      parameters:
        - name: user_id
          in: path
          required: true
          description: User identifier
          schema:
            type: string
            example: chatgpt_user_abc123
      responses:
        "200":
          description: Full progress summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProgress"
  /api/v1/progress/{user_id}/chapter/{chapter_id}:
    put:
      operationId: update_progress
      summary: Update chapter progress
      description: >
        Updates a user's progress for a specific chapter. Automatically
        recalculates the study streak. Call this when a student marks a
        chapter as complete or reaches a milestone completion percentage.
      parameters:
        - name: user_id
          in: path
          required: true
          description: User identifier
          schema:
            type: string
            example: chatgpt_user_abc123
        - name: chapter_id
          in: path
          required: true
          description: Chapter identifier
          schema:
            type: string
            example: chapter-01
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProgressUpdate"
      responses:
        "200":
          description: Updated chapter progress row
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChapterProgressOut"
  /api/v1/progress/{user_id}/streak:
    get:
      operationId: get_streak
      summary: Get user streak information
      description: >
        Returns streak-only data: current streak, longest streak,
        last activity date, and total study days. Faster than
        get_progress when only streak data is needed.
      parameters:
        - name: user_id
          in: path
          required: true
          description: User identifier
          schema:
            type: string
            example: chatgpt_user_abc123
      responses:
        "200":
          description: Streak information
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StreakInfo"

  # ── 6.5 Search API ───────────────────────────────────────────────────────────
  /api/v1/search:
    get:
      operationId: search_content
      summary: Search course content
      description: >
        Keyword search across chapter titles, descriptions, tags, and full
        chapter text. Only returns chapters the requesting user has access to
        based on their tier. Each result includes a 150-character excerpt
        showing where the match was found.
      parameters:
        - name: q
          in: query
          required: true
          description: Search keyword or phrase (e.g. "MCP protocol")
          schema:
            type: string
            minLength: 1
            example: MCP protocol
        - name: user_id
          in: query
          required: false
          description: User identifier for access-filtered results
          schema:
            type: string
            example: chatgpt_user_abc123
      responses:
        "200":
          description: List of matching chapter sections with excerpts
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SearchResponse"

  # ── 6.6 Access Control APIs ──────────────────────────────────────────────────
...
  /api/v1/access/users/{user_id}/tier:
    get:
      operationId: get_user_tier
      summary: Get user subscription tier
      description: >
        Returns the subscription tier for a user_id.
        Returns 'free' if the user is not registered in the system.
        Possible values: free | premium | pro | team
      parameters:
        - name: user_id
          in: path
          required: true
          description: User identifier
          schema:
            type: string
            example: chatgpt_user_abc123
      responses:
        "200":
          description: User tier
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TierResponse"

# ─────────────────────────────────────────────────────────────────────────────
# COMPONENTS / SCHEMAS
# ─────────────────────────────────────────────────────────────────────────────
components:
  schemas:
...
    QuizIdResponse:
      type: object
      properties:
        chapter_id:
          type: string
          example: chapter-01
        quiz_id:
          type: string
          example: quiz-01
      required: [chapter_id, quiz_id]
    QuizQuestion:
      type: object
      properties:
        question_id:
          type: string
          example: q1
        question:
          type: string
          example: What does FTE stand for?
        options:
          type: object
          additionalProperties:
            type: string
          example:
            A: Full Task Execution
            B: Full-Time Equivalent
            C: Functional Token Engine
            D: Fast Training Evaluator
        type:
          type: string
          example: mcq
      required: [question_id, question, options, type]
    QuizOut:
      type: object
      description: Quiz with questions but WITHOUT correct answers
      properties:
        quiz_id:
          type: string
          example: quiz-01
        chapter_id:
          type: string
          example: chapter-01
        title:
          type: string
          example: "Chapter 1: Introduction to AI Agents and Digital FTEs"
        pass_threshold:
          type: integer
          example: 70
        questions:
          type: array
          items:
            $ref: "#/components/schemas/QuizQuestion"
      required: [quiz_id, chapter_id, title, pass_threshold, questions]
    AnswerItem:
      type: object
      properties:
        question_id:
          type: string
          example: q1
        answer:
          type: string
          enum: [A, B, C, D]
          example: B
      required: [question_id, answer]
    QuizSubmission:
      type: object
      properties:
        user_id:
          type: string
          example: chatgpt_user_abc123
        answers:
          type: array
          items:
            $ref: "#/components/schemas/AnswerItem"
          minItems: 1
      required: [user_id, answers]
    QuestionFeedback:
      type: object
      properties:
        question_id:
          type: string
          example: q1
        correct:
          type: boolean
          example: true
        correct_answer:
          type: string
          example: B
        explanation:
          type: string
          nullable: true
          description: Explanation shown only for incorrect answers
      required: [question_id, correct, correct_answer]
    QuizResult:
      type: object
      properties:
        quiz_id:
          type: string
          example: quiz-01
        user_id:
          type: string
          example: chatgpt_user_abc123
        score:
          type: integer
          minimum: 0
          maximum: 100
          example: 80
        total:
          type: integer
          example: 5
        correct:
          type: integer
          example: 4
        passed:
          type: boolean
          example: true
        feedback:
          type: array
          items:
            $ref: "#/components/schemas/QuestionFeedback"
      required: [quiz_id, user_id, score, total, correct, passed, feedback]

    # ── Progress ──────────────────────────────────────────────────────────────
    QuizScore:
      type: object
      properties:
        quiz_id:
          type: string
          example: quiz-01
        score:
          type: integer
          example: 80
        passed:
          type: boolean
          example: true
        attempted_at:
          type: string
          format: date-time
          nullable: true
      required: [quiz_id, score, passed]
    UserProgress:
      type: object
      properties:
        user_id:
          type: string
          example: chatgpt_user_abc123
        completed_chapters:
          type: array
          items:
            type: string
          example: ["chapter-01", "chapter-02"]
        completion_percentage:
          type: integer
          minimum: 0
          maximum: 100
          example: 33
        current_streak:
          type: integer
          example: 5
        longest_streak:
          type: integer
          example: 7
        total_study_days:
          type: integer
          example: 12
        quiz_scores:
          type: array
          items:
            $ref: "#/components/schemas/QuizScore"
      required:
        - user_id
        - completed_chapters
        - completion_percentage
        - current_streak
        - longest_streak
        - total_study_days
        - quiz_scores
    ProgressUpdate:
      type: object
      properties:
        completed:
          type: boolean
          example: true
        completion_percentage:
          type: integer
          minimum: 0
          maximum: 100
          example: 100
      required: [completed, completion_percentage]
    ChapterProgressOut:
      type: object
      properties:
        user_id:
          type: string
        chapter_id:
          type: string
        completed:
          type: boolean
        completion_percentage:
          type: integer
        last_read_at:
          type: string
          format: date-time
          nullable: true
        updated_at:
          type: string
          format: date-time
          nullable: true
      required: [user_id, chapter_id, completed, completion_percentage]
    StreakInfo:
      type: object
      properties:
        user_id:
          type: string
          example: chatgpt_user_abc123
        current_streak:
          type: integer
          example: 5
        longest_streak:
          type: integer
          example: 7
        last_activity_date:
          type: string
          format: date
          nullable: true
          example: "2026-04-11"
        total_study_days:
          type: integer
          example: 12
      required: [user_id, current_streak, longest_streak, total_study_days]

    # ── Search ────────────────────────────────────────────────────────────────
    SearchResult:
      type: object
      properties:
        chapter_id:
          type: string
          example: chapter-03
        title:
          type: string
          example: MCP and Agent Skills
        tier_required:
          type: string
          enum: [free, premium, pro]
        order_index:
          type: integer
          example: 3
        excerpt:
          type: string
          description: 150-character excerpt around the match location
          example: "...MCP stands for Model Context Protocol — it provides tools and domain knowledge to agents..."
      required: [chapter_id, title, tier_required, order_index, excerpt]
    SearchResponse:
      type: object
      properties:
        query:
          type: string
          example: MCP protocol
        user_id:
          type: string
          example: chatgpt_user_abc123
        total:
          type: integer
          example: 2
        results:
          type: array
          items:
            $ref: "#/components/schemas/SearchResult"
      required: [query, user_id, total, results]

    # ── Access ────────────────────────────────────────────────────────────────
...
    TierResponse:
      type: object
      properties:
        user_id:
          type: string
          example: chatgpt_user_abc123
        tier:
          type: string
          enum: [free, premium, pro, team]
          example: free
      required: [user_id, tier]

    # ── Errors ────────────────────────────────────────────────────────────────
    ErrorDetail:
      type: object
      properties:
        detail:
          type: string
          example: Chapter 'chapter-99' not found
      required: [detail]
...